change: 增加设置快捷入口的activity

+++ HDLocker/src/cn/zmdx/kaka/fast/locker/shortcut/ShortcutManager.java
@@ -7,19 +7,17 @@ import java.util.List;
 
 import android.content.Context;
 import android.content.Intent;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.LayoutInflater;
-import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
 import cn.zmdx.kaka.fast.locker.LockScreenManager;
 import cn.zmdx.kaka.fast.locker.R;
 import cn.zmdx.kaka.fast.locker.database.ShortcutModel;
+import cn.zmdx.kaka.fast.locker.settings.ShortcutSettingsActivity;
 import cn.zmdx.kaka.fast.locker.utils.BaseInfoHelper;
-import cn.zmdx.kaka.fast.locker.widget.RippleView;
-import cn.zmdx.kaka.fast.locker.widget.RippleView.Callback;
 import cn.zmdx.kaka.fast.locker.widget.dragdropgridview.DragDropGrid;
-import cn.zmdx.kaka.fast.locker.widget.dragdropgridview.DragDropGrid.OnItemClickListener;
 
 public class ShortcutManager {
 
@@ -35,7 +33,7 @@ public class ShortcutManager {
         mContext = context;
         mInflater = LayoutInflater.from(context);
         // test
-        initTestData();
+        // initTestData();
     }
 
     // test method
@@ -72,15 +70,15 @@ public class ShortcutManager {
         ai2.setShortcut(true);
         data.add(ai2);
 
-        AppInfo ai3 = new AppInfo();
-        ai3.setAppName("拨号");
-        String pkgName3 = "com.tencent.pb";
-        ai3.setPkgName(pkgName3);
-        ai3.setDefaultIcon(AppInfo.getIconByPkgName(mContext, pkgName3));
-        ai3.setDisguise(false);
-        ai3.setPosition(2);
-        ai3.setShortcut(true);
-        data.add(ai3);
+        // AppInfo ai3 = new AppInfo();
+        // ai3.setAppName("拨号");
+        // String pkgName3 = "com.tencent.pb";
+        // ai3.setPkgName(pkgName3);
+        // ai3.setDefaultIcon(AppInfo.getIconByPkgName(mContext, pkgName3));
+        // ai3.setDisguise(false);
+        // ai3.setPosition(2);
+        // ai3.setShortcut(true);
+        // data.add(ai3);
 
         AppInfo ai4 = new AppInfo();
         ai4.setAppName("Chrome");
@@ -88,7 +86,7 @@ public class ShortcutManager {
         ai4.setPkgName(pkgName4);
         ai4.setDefaultIcon(AppInfo.getIconByPkgName(mContext, pkgName4));
         ai4.setDisguise(false);
-        ai4.setPosition(6);
+        ai4.setPosition(0);
         ai4.setShortcut(true);
         data.add(ai4);
 
@@ -113,50 +111,50 @@ public class ShortcutManager {
 
     public View createShortcutAppsView() {
         View view = mInflater.inflate(R.layout.tool_box_layout, null);
+        // test
+        view.findViewById(R.id.settings).setOnClickListener(new View.OnClickListener() {
+
+            @Override
+            public void onClick(View v) {
+                Intent in = new Intent(mContext, ShortcutSettingsActivity.class);
+                in.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                mContext.startActivity(in);
+                LockScreenManager.getInstance().unLock();
+            }
+        });
         mGridView = (DragDropGrid) view.findViewById(R.id.gridView);
         ViewGroup.LayoutParams lp = mGridView.getLayoutParams();
         lp.height = BaseInfoHelper.getRealWidth(mContext) / 3 * 2;
         List<AppInfo> data = getShortcutInfo();
         final ShortcutAppAdapter adapter = new ShortcutAppAdapter(mContext, mGridView, data);
         mGridView.setAdapter(adapter);
-        mGridView.setOnItemClickListener(new OnItemClickListener() {
-
-            @Override
-            public void onItemClick(View clickedView, Object data, MotionEvent event) {
-                if (data instanceof AppInfo) {
-                    final AppInfo ai = (AppInfo) data;
-                    if (clickedView instanceof RippleView) {
-                        RippleView rv = (RippleView) clickedView;
-                        rv.animateRipple(event, new Callback() {
-
-                            @Override
-                            public void onFinish(View v) {
-                                startTarget(ai);
-                            }
-                        });
-                        return;
-                    }
-                    startTarget(ai);
-                }
-            }
-        });
         return view;
     }
 
-    private void startTarget(AppInfo ai) {
-        Intent intent = mContext.getPackageManager().getLaunchIntentForPackage(ai.getPkgName());
-        if (intent != null) {
-            mContext.startActivity(intent);
-            LockScreenManager.getInstance().unLock();
-        }
-    }
-
+    /**
+     * 获取6个快捷入口的集合，如果入口不足6个，会用假的AppInfo对象填充，返回的集合根据AppInfo的position升序排列
+     * 
+     * @return
+     */
     @SuppressWarnings("unchecked")
     public List<AppInfo> getShortcutInfo() {
         List<AppInfo> data = ShortcutModel.getInstance(mContext).queryAll();
-        if (data != null && data.size() > 0) {
-            Collections.sort(data);
+        List<AppInfo> tmpList = new ArrayList<AppInfo>();
+        for (int pos = 0; pos < 6; pos++) {
+            boolean posExsit = false;
+            for (int i = 0; i < data.size(); i++) {
+                if (pos == data.get(i).getPosition()) {
+                    posExsit = true;
+                }
+            }
+            if (!posExsit) {
+                AppInfo ai = new AppInfo();
+                ai.setPosition(pos);
+                tmpList.add(ai);
+            }
         }
+        data.addAll(tmpList);
+        Collections.sort(data);
         return data;
     }
 
@@ -164,10 +162,15 @@ public class ShortcutManager {
      * @param data
      */
     public void saveShortcutInfo(List<AppInfo> data) {
+        if (data == null || data.size() <= 0) {
+            return;
+        }
         ShortcutModel sm = ShortcutModel.getInstance(mContext);
         sm.deleteAll();
         for (AppInfo ai : data) {
-            boolean result = sm.insert(ai);
+            if (!TextUtils.isEmpty(ai.getPkgName())) {
+                boolean result = sm.insert(ai);
+            }
         }
     }
 }
